Split fetching data for the national-highseas table's filters into a separate hook for readability
